requirements.txt

# Python dependencies
Flask==2.2.2
gunicorn==20.1.0
google-cloud-firestore==2.7.2
